vuart_config: separate vuart config from vm config file
Not add mmio vuart support now.

Tracked-On: #8805
Acked-by: Wang, Yu1 <yu1.wang@intel.com>
