+++ hypervisor/include/arch/x86/asm/vm_config.h
@@ -16,6 +16,7 @@
 #include <asm/sgx.h>
 #include <acrn_hv_defs.h>
 #include <schedule.h>
+#include <vuart_config.h>
 
 #define AFFINITY_CPU(n)		(1UL << (n))
 #define MAX_VCPUS_PER_VM	MAX_PCPU_NUM
@@ -85,32 +86,6 @@ struct acrn_vm_mem_config {
 	uint64_t region_num;
 	struct vm_hpa_regions  *host_regions;
 };
-
-struct target_vuart {
-	uint8_t vm_id;		/* target VM id */
-	uint8_t vuart_id;	/* target vuart index in a VM */
-};
-
-enum vuart_type {
-	VUART_LEGACY_PIO = 0,	/* legacy PIO vuart */
-	VUART_PCI,		/* PCI vuart, may removed */
-};
-
-union vuart_addr {
-	uint16_t port_base;		/* addr for legacy type */
-	struct {			/* addr for pci type */
-		uint8_t f : 3;		/* BITs 0-2 */
-		uint8_t d : 5;		/* BITs 3-7 */
-		uint8_t b;		/* BITs 8-15 */
-	} bdf;
-};
-
-struct vuart_config {
-	enum vuart_type type;		/* legacy PIO or PCI  */
-	union vuart_addr addr;		/* port addr if in legacy type, or bdf addr if in pci type */
-	uint16_t irq;
-	struct target_vuart t_vuart;	/* target vuart */
-} __aligned(8);
 
 enum os_kernel_type {
 	KERNEL_BZIMAGE = 1,
